+++ src/common/state/common-ids.ts
@@ -1,4 +1,6 @@
 export const CODE_EDITOR = "CodeEditor";
+export const TEXT_EDITOR = "TextEditor";
+export const BIN_VIEWER = "BinViewer";
 export const DISASSEMBLY_PANEL_ID = "$disassembly";
 export const DISASSEMBLY_EDITOR = "Disassembly";
 export const MEMORY_PANEL_ID = "$memory";
@@ -23,4 +25,5 @@ export const NPL_EDITOR = "NplEditor";
 export const NXI_EDITOR = "NxiEditor";
 export const SPR_EDITOR = "SprEditor";
 export const VID_VIEWER = "VidViewer";
+export const IMAGE_VIEWER = "ImageViewer";
 export const SCRIPT_OUTPUT_VIEWER = "ScriptOutputViewer";